+++ 1.21.6/src/main/java/com/bleudev/ppl_utils/mixin/client/DebugHudMixin.java
@@ -1,0 +1,28 @@
+package com.bleudev.ppl_utils.mixin.client;
+
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.gui.hud.DebugHud;
+import org.jetbrains.annotations.NotNull;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.bleudev.ppl_utils.util.LangUtils.round;
+
+@Mixin(DebugHud.class)
+public class DebugHudMixin {
+    @Inject(method = "getLeftText", at = @At("RETURN"), cancellable = true)
+    private void addWorldBorderDebugHudEntry(@NotNull CallbackInfoReturnable<List<String>> cir) {
+        var l = new ArrayList<>(cir.getReturnValue());
+
+        MinecraftClient client = MinecraftClient.getInstance();
+        if (client.world != null)
+            l.add("World border: " + round(client.world.getWorldBorder().getSize(), 3));
+
+        cir.setReturnValue(l);
+    }
+}