+++ 1.21.9/src/main/java/com/bleudev/ppl_utils/client/custom/WorldBorderDebugHudEntry.java
@@ -7,11 +7,13 @@ import net.minecraft.world.World;
 import net.minecraft.world.chunk.WorldChunk;
 import org.jetbrains.annotations.Nullable;
 
+import static com.bleudev.ppl_utils.util.LangUtils.round;
+
 public class WorldBorderDebugHudEntry implements DebugHudEntry {
     @Override
     public void render(DebugHudLines lines, @Nullable World world, @Nullable WorldChunk clientChunk, @Nullable WorldChunk chunk) {
         MinecraftClient client = MinecraftClient.getInstance();
         if (client.world != null)
-            lines.addLine("World border: " + client.world.getWorldBorder().getSize());
+            lines.addLine("World border: " + round(client.world.getWorldBorder().getSize(), 3));
     }
 }