+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+# 2.1
+## Minecraft updates
+[+] Update for 1.21.6-1.21.8
+## Fixes
+[/] World border precision changed to `10^3`\
+[-] Deleted unused `com.bleudev.ppl_utils.client.impl.LobbyButtonText` class
+
 # 2.0
 ## Breaking changes
 [/] Rewrite from Kotlin to Java
 ## New features
 [+] World border size debug hud entry (thanks `wiihxhx` for idea)\
-[+] Beta mode
+[+] Beta mode\
 [+] Set logo, description, links, badges and other metadata
 ## Updates
 [+] Updated lobby button rendering